👯 Split API into Guardian and Mediator modes (#74)
Guardian mode endpoints:
- /guardian/*
- /key
  - /auxiliary/generate
  - /election/generate
- /tally
  - /decrypt-share
- /ping

Mediator mode endtpoints:
- /ballot/*
- /election/*
- /key/election/combine
- /tally
  - /
  - /append
  - /decrypt
- /ping

The mode is determined on startup by the `API_MODE` environment variable, which can be set to either `guardian` or `mediator` (default).

The docker image accepts this environment variable if provided, and a docker-compose file has been added for easy local execution of the image in both guardian and mediator modes.
