+++ app/api/v1/api.py
@@ -1,11 +1,19 @@
 from fastapi import APIRouter
-
-from app.api.v1.endpoints import ballot, election, guardian, key, ping, tally
+from app.core.config import settings, ApiMode
+from . import common
 
 api_router = APIRouter()
-api_router.include_router(ballot.router, prefix="/ballot", tags=["ballot"])
-api_router.include_router(election.router, prefix="/election", tags=["election"])
-api_router.include_router(guardian.router, prefix="/guardian", tags=["guardian"])
-api_router.include_router(key.router, prefix="/key", tags=["key"])
-api_router.include_router(ping.router, prefix="/ping", tags=["ping"])
-api_router.include_router(tally.router, prefix="/tally", tags=["tally"])
+
+if settings.API_MODE == ApiMode.guardian:
+    from . import guardian
+
+    api_router.include_router(guardian.router)
+elif settings.API_MODE == ApiMode.mediator:
+    from . import mediator
+
+    api_router.include_router(mediator.router)
+else:
+    raise ValueError(f"Unknown API mode: {settings.API_MODE}")
+
+
+api_router.include_router(common.router)
