+++ app/api/v1/common/__init__.py
@@ -1,0 +1,1 @@
+from .routes import router
