+++ app/api/v1/common/routes.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+from . import ping
+
+router = APIRouter()
+
+router.include_router(ping.router, prefix="/ping")
