--- app/api/v1/endpoints/ping.py
@@ -1,13 +1,0 @@
-from typing import Any
-
-from fastapi import APIRouter
-
-router = APIRouter()
-
-
-@router.get("", response_model=str)
-def ping() -> Any:
-    """
-    Ensure API can be pinged
-    """
-    return "pong"
