+++ app/api/v1/mediator/__init__.py
@@ -1,0 +1,1 @@
+from .routes import router
