+++ app/api/v1/tags.py
@@ -1,1 +1,6 @@
-GUARDIAN_ONLY = "Guardian Only"
+CONFIGURE_ELECTION = "Configure Election"
+KEY_CEREMONY = "Key Ceremony"
+ENCRYPT_BALLOTS = "Encrypt Ballots"
+CAST_AND_SPOIL = "Cast and Spoil"
+TALLY = "Tally"
+UTILITY = "Utility"
