+++ app/core/config.py
@@ -1,12 +1,19 @@
+from enum import Enum
 from pydantic import AnyHttpUrl, BaseSettings
-from typing import cast, List
+from pydantic.fields import Field
+from typing import List
+
+
+class ApiMode(str, Enum):
+    guardian = "guardian"
+    mediator = "mediator"
 
 
 class Settings(BaseSettings):
+    API_MODE: ApiMode = ApiMode.mediator
     API_V1_STR: str = "/api/v1"
-    BACKEND_CORS_ORIGINS: List[AnyHttpUrl] = cast(
-        List[AnyHttpUrl],
-        ["http://localhost", "http://localhost:8080", "https://localhost"],
+    BACKEND_CORS_ORIGINS: List[AnyHttpUrl] = Field(
+        default=["http://localhost", "http://localhost:8080", "https://localhost"]
     )
     PROJECT_NAME: str = "electionguard-web-api"
