+++ app/main.py
@@ -1,12 +1,17 @@
 from fastapi import FastAPI
+from logging import getLogger
 from starlette.middleware.cors import CORSMiddleware
 
 from app.api.v1.api import api_router
 from app.core.config import settings
 
+logger = getLogger("app")
+
 app = FastAPI(
     title=settings.PROJECT_NAME, openapi_url=f"{settings.API_V1_STR}/openapi.json"
 )
+
+logger.error(f"Starting API in {settings.API_MODE} mode")
 
 # Set all CORS enabled origins
 if settings.BACKEND_CORS_ORIGINS:
@@ -28,5 +33,18 @@ if __name__ == "__main__":
     # `PYTHONPATH=. pipenv run python ./app/main.py`
     # See the VSCode launch configuration for detail.
     import uvicorn
+    import argparse
 
-    uvicorn.run(app, host="0.0.0.0", port=8000)
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+    parser.add_argument(
+        "-p",
+        "--port",
+        default=8000,
+        type=int,
+        help="The port to listen on",
+    )
+    args = parser.parse_args()
+
+    uvicorn.run(app, host="0.0.0.0", port=args.port)
